Finished refactoring and added the entity parameter to crossbow speed
