--- fabric-dehardcoding-api-v0/build.gradle
@@ -1,6 +1,0 @@
-archivesBaseName = "fabric-extensibility-api-v1"
-version = getSubprojectVersion(project, "1.0.0")
-
-dependencies {
-	compile project(path: ':fabric-api-base', configuration: 'dev')
-}
