--- fabric-dehardcoding-api-v0/src/main/resources/fabric-extensibility-api-v1.mixins.json
@@ -1,23 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.fabricmc.fabric.mixin.dehardcoding",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "CrossbowItemMixin",
-    "EntityCrossbowUseMixin",
-    "EntityBowUseMixin",
-    "AbstractSkeletonEntityMixin",
-    "ProjectileUtilMixin",
-    "BowItemMixin",
-    "ItemMixin"
-  ],
-  "client": [
-    "client.HeldItemRendererMixin",
-    "client.PlayerEntityRendererMixin",
-    "client.SkeletonEntityModelMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
