+++ fabric-extensibility-api-v1/src/main/resources/assets/fabric-extensibility-api-v1/icon.png
