+++ fabric-extensibility-api-v1/src/main/resources/fabric-extensibility-api-v1.accesswidener
@@ -1,0 +1,3 @@
+accessWidener	v1	named
+
+accessible    method    net/minecraft/item/CrossbowItem    getSpeed    (Lnet/minecraft/item/ItemStack;)F
